requirements.txt
duckdb
pandas
pyarrow
pyiceberg
python-dotenv